Google Cloud Function/requirements.txt
python-dateutil~=2.8.2
google-cloud-storage==2.4.0
